requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
spacy==3.6.1
PyMuPDF==1.22.5
python-docx==0.8.11
extract-msg==0.29.0
requests==2.31.0
python-dotenv==1.0.0
email-validator==1.3.1
numpy==1.26.4
sentence-transformers==2.2.2
faiss-cpu==1.7.4
google-generativeai==0.11.0
